Rename conversation.gpt4all package to conversation.offline
