+++ src/khoj/processor/conversation/offline/chat_model.py
